Consolidate Performance events into PerformanceTimer.
Use an integer for durations.

+++ types/kotlin/Screen.kt
@@ -4,7 +4,11 @@ package quicktype
  * Triggered when the user changed screen
  */
 data class Screen (
-    val durationMS: Double? = null,
+    /**
+     * How long the screen was displayed for in milliseconds.
+     */
+    val durationMS: Long? = null,
+
     val eventName: EventName,
     val screenName: ScreenName
 )
